requirements.txt
requests
mss
pyautogui
Pillow